[am335x] Fix HDMI suport for Beaglebone Black
Fallback to Linux video interface bindings introduced in r313068 worked
with then current DTS but that DTS turned out to be not conformant to
the the bindings spec. DTS import in r314854 fixed the conformancy but
broke the functionality. This commit syncs up functionality to the actual
spec.

Reported by:	manu@


git-svn-id: https://svn.freebsd.org/base/head@316306 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f