MFC r306041: Always pass -m to ld for converting binary files to ELF
This is in preparation for linking with LLVM's lld, which does not have
a compiled-in default output emulation. lld requires that it is
specified via the -m option, or obtained from the object file(s) being
linked.

This will also allow all build targets to share a common linker binary.

Sponsored by:	The FreeBSD Foundation


git-svn-id: https://svn.freebsd.org/base/stable/11@316469 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f